improving the API

+++ Controllers/HomeController.cs
@@ -8,31 +8,33 @@ namespace Todo.Controllers;
 public class HomeController : ControllerBase
 {
     [HttpGet("/")]
-    public List<TodoModel> Get([FromServices] AppDbContext context)
-    {
-        return context.TodoModels.ToList();
-    }
+    public IActionResult Get([FromServices] AppDbContext context) 
+        => Ok(context.TodoModels.ToList());
 
     [HttpGet("/{id:int}")]
-    public TodoModel GetById([FromRoute] int id, [FromServices] AppDbContext context)
+    public IActionResult GetById([FromRoute] int id, [FromServices] AppDbContext context)
     {
-        return context.TodoModels.FirstOrDefault(x => x.Id == id);
+        var todos = context.TodoModels.FirstOrDefault(x => x.Id == id);
+        if (todos == null)
+            return NotFound();
+
+        return Ok(todos);
     }
 
     [HttpPost("/")]
-    public TodoModel Post([FromBody] TodoModel todo, [FromServices] AppDbContext context)
+    public IActionResult Post([FromBody] TodoModel todo, [FromServices] AppDbContext context)
     {
         context.TodoModels.Add(todo);
         context.SaveChanges();
-        return todo;
+        return Created($"/{todo.Id}",todo);
     }
     
     [HttpPut("/{id:int}")]
-    public TodoModel Put([FromRoute]int id, [FromBody] TodoModel todo, [FromServices] AppDbContext context)
+    public IActionResult Put([FromRoute]int id, [FromBody] TodoModel todo, [FromServices] AppDbContext context)
     {
         var model = context.TodoModels.FirstOrDefault(x => x.Id ==  id);
         if (model == null)
-            return todo;
+            return NotFound();
         
         // Comparar os valores que vem da tabela com que veio do banco 
         model.Title = todo.Title;
@@ -40,15 +42,18 @@ public class HomeController : ControllerBase
 
         context.TodoModels.Update(model);
         context.SaveChanges();
-        return model;
+        return Ok(model);
     }
     
     [HttpDelete("/{id:int}")]
-    public TodoModel Delete([FromRoute]int id, [FromServices] AppDbContext context)
+    public IActionResult Delete([FromRoute]int id, [FromServices] AppDbContext context)
     {
         var model = context.TodoModels.FirstOrDefault(x => x.Id ==  id);
+        if (model == null)
+            return NotFound();
+        
         context.TodoModels.Remove(model);
         context.SaveChanges();
-        return model;
+        return Ok(model);
     }
 }
